optimize product detail
