+++ ECommerce.WebApp/client/src/_pages/product-detail/_components/dialog/prod-gallery.tsx
@@ -1,7 +1,6 @@
 import Dialog from "@mui/material/Dialog";
 
 import DialogContent from "@mui/material/DialogContent";
-import { imagesSlide } from "../silde/product-detail-slide.component";
 import { MuiIcon } from "src/_shares/_components";
 
 export default function GalleryDialog(props: any) {
