+++ ECommerce.WebApp/client/src/_pages/product-detail/_components/layout/prod-suggestion.tsx
