+++ ECommerce.WebApp/client/src/_pages/product-detail/_components/slide/prod-real-images.tsx
