+++ ECommerce.WebApp/client/src/_pages/product-detail/_components/tab/detail-info.tsx
@@ -1,0 +1,64 @@
+export default function DetailInfo(props: any) {
+    return (
+        <div className="product-detail">
+            <ul className="detail-list flex mb-4 flex-wrap justify-between">
+                <li className="detail-item">
+                    <span>
+                        <b>Chất liệu</b>
+                    </span>
+                    <span>Thủy tinh cao cấp</span>
+                </li>
+                <li className="detail-item">
+                    <span>
+                        <b>Chất liệu</b>
+                    </span>
+                    <span>Thủy tinh cao cấp</span>
+                </li>
+                <li className="detail-item">
+                    <span>
+                        <b>Chất liệu</b>
+                    </span>
+                    <span>Thủy tinh cao cấp</span>
+                </li>
+                <li className="detail-item">
+                    <span>
+                        <b>Chất liệu</b>
+                    </span>
+                    <span>Thủy tinh cao cấp</span>
+                </li>
+                <li className="detail-item">
+                    <span>
+                        <b>Chất liệu</b>
+                    </span>
+                    <span>Thủy tinh cao cấp</span>
+                </li>
+                <li className="detail-item">
+                    <span>
+                        <b>Chất liệu</b>
+                    </span>
+                    <span>Thủy tinh cao cấp</span>
+                </li>
+                <li className="detail-item">
+                    <span>
+                        <b>Chất liệu</b>
+                    </span>
+                    <span>Thủy tinh cao cấp</span>
+                </li>
+                <li className="detail-item">
+                    <span>
+                        <b>Chất liệu</b>
+                    </span>
+                    <span>Thủy tinh cao cấp</span>
+                </li>
+                <li className="detail-item">
+                    <span>
+                        <b>Chất liệu</b>
+                    </span>
+                    <span>Thủy tinh cao cấp</span>
+                </li>
+            </ul>
+            <div className="title">Mô tả ngắn</div>
+            <div className="description"></div>
+        </div>
+    )
+}
