+++ ECommerce.WebApp/client/src/_pages/product-detail/_interfaces/comment.interface.ts
@@ -1,3 +1,4 @@
 export interface IUserComment {
+    id: string;
     type: 'post' | 'comment' | 'reply'
 }
