+++ ECommerce.WebApp/client/src/_shares/_components/input/number.tsx
@@ -1,0 +1,9 @@
+interface INumberProps extends React.HTMLProps<HTMLInputElement> {
+
+}
+
+export default function NumberInput(props: INumberProps) {
+    return (
+        <input type="number" {...props} />
+    )
+}
